refactor(audio)!: move spatial descriptor state from Sound to Voice/PlayOptions
Sound becomes a stateless asset descriptor; position, velocity,
distanceModel, refDistance, maxDistance, and rolloffFactor now live
entirely on the per-playback Voice (settable live, or seeded via
PlayOptions at play() time). Fixes the redundant sound.position +
voice.position duplication visible in examples/spatial-audio/moving-source.ts.

Also re-exports DistanceModel from Playable.ts (its new home) at the
package root, and updates test/audio/audio-manager-update.test.ts's
three call sites that still exercised the removed Sound.position setter.
