+++ src/audio/Playable.ts
@@ -85,6 +85,19 @@ export interface RatePitched {
   detune: number;
 }
 
+/**
+ * Distance-attenuation model used by spatial sounds.
+ *
+ * Mirrors Web Audio's `PannerNode.distanceModel`:
+ * - `'linear'` — `v = 1 - rolloffFactor * (d - refDistance) / (maxDistance - refDistance)`,
+ *   clamped to [0, 1]. Reaches silence at `maxDistance`.
+ * - `'inverse'` — `v = refDistance / (refDistance + rolloffFactor * (d - refDistance))`.
+ *   Physically realistic; never reaches absolute silence.
+ * - `'exponential'` — `v = (d / refDistance) ^ -rolloffFactor`. Steepest near
+ *   the listener; useful for very intimate sources.
+ */
+export type DistanceModel = 'linear' | 'inverse' | 'exponential';
+
 /** A voice that can be positioned in 2D space and optionally track a node. */
 export interface Spatializable {
   /** World-space position of the source, or `null` when not spatialized. */
@@ -97,6 +110,14 @@ export interface Spatializable {
    * {@link Spatializable.position}.
    */
   follow(node: SceneNode | null): void;
+  /** Distance-attenuation model. Default `'linear'`. */
+  distanceModel: DistanceModel;
+  /** Distance below which volume is at full strength. Default `50`. */
+  refDistance: number;
+  /** For the `'linear'` model: distance at which volume reaches zero. Default `1000`. */
+  maxDistance: number;
+  /** Falloff rate. Higher = steeper attenuation. Default `1`. */
+  rolloffFactor: number;
 }
 
 /**
@@ -117,6 +138,16 @@ export interface PlayOptions {
   time?: number;
   /** Start muted (volume 0). */
   muted?: boolean;
+  /** Initial spatial position — equivalent to setting `voice.position` right after play. */
+  position?: { x: number; y: number } | Vector;
+  /** Initial distance-attenuation model. Default `'linear'`. */
+  distanceModel?: DistanceModel;
+  /** Initial reference distance. Default `50`. */
+  refDistance?: number;
+  /** Initial max distance (`'linear'` model only). Default `1000`. */
+  maxDistance?: number;
+  /** Initial rolloff factor. Default `1`. */
+  rolloffFactor?: number;
 }
 
 /**
