+++ src/audio/index.ts
@@ -16,8 +16,8 @@ export type { EnvelopeOptions } from './Envelope';
 export { Envelope } from './Envelope';
 export type { HighpassFilterOptions, LowpassFilterOptions } from './filters';
 export { HighpassFilter, LowpassFilter } from './filters';
-export type { Loopable, Pausable, Playable, PlayOptions, RatePitched, Seekable, Spatializable, Voice } from './Playable';
-export type { AudioSpriteClip, DistanceModel, SoundOptions, SoundPlayOptions } from './Sound';
+export type { DistanceModel, Loopable, Pausable, Playable, PlayOptions, RatePitched, Seekable, Spatializable, Voice } from './Playable';
+export type { AudioSpriteClip, SoundOptions, SoundPlayOptions } from './Sound';
 export { Sound, SoundPoolStrategy } from './Sound';
 export type { SpatialSmoothingSettings } from './spatial-smoothing';
 export { WorkletEffect } from './WorkletEffect';
